Oof, refactored the display panel.... now we can add the cool new stuff
