+++ UI/Dialogs/ConfigWizard.cs
@@ -10,6 +10,7 @@ using MobiFlight;
 using MobiFlight.Base;
 using MobiFlight.UI.Forms;
 using MobiFlight.UI.Panels.Config;
+using MobiFlight.UI.Panels.OutputWizard;
 
 namespace MobiFlight.UI.Dialogs
 {
@@ -20,8 +21,6 @@ namespace MobiFlight.UI.Dialogs
         static int lastTabActive = 0;
 
         ExecutionManager _execManager = null;
-        int displayPanelHeight = -1;
-        List<UserControl> displayPanels = new List<UserControl>();
         OutputConfigItem config = null;
         OutputConfigItem originalConfig = null;
         ErrorProvider errorProvider = new ErrorProvider();
@@ -31,15 +30,6 @@ namespace MobiFlight.UI.Dialogs
         Dictionary<String, String> arcazeFirmware = new Dictionary<String, String>();
         Dictionary<string, ArcazeModuleSettings> moduleSettings;
 #endif
-
-        Panels.DisplayPinPanel              displayPinPanel             = new Panels.DisplayPinPanel();
-        Panels.DisplayBcdPanel              displayBcdPanel             = new Panels.DisplayBcdPanel();
-        Panels.DisplayLedDisplayPanel       displayLedDisplayPanel      = new Panels.DisplayLedDisplayPanel();
-        Panels.DisplayNothingSelectedPanel  displayNothingSelectedPanel = new Panels.DisplayNothingSelectedPanel();
-        Panels.LCDDisplayPanel              displayLcdDisplayPanel      = new Panels.LCDDisplayPanel();
-        Panels.ServoPanel                   servoPanel                  = new Panels.ServoPanel();
-        Panels.StepperPanel                 stepperPanel                = new Panels.StepperPanel();
-        Panels.DisplayShiftRegisterPanel    displayShiftRegisterPanel   = new Panels.DisplayShiftRegisterPanel();
 
         public ConfigWizard( ExecutionManager mainForm, 
                              OutputConfigItem cfg,
@@ -58,7 +48,14 @@ namespace MobiFlight.UI.Dialogs
             initWithoutArcazeCache();
 #endif
             preconditionPanel.preparePreconditionPanel(dataSetConfig, filterGuid);
-            initConfigRefDropDowns(dataSetConfig, filterGuid);          
+            initConfigRefDropDowns(dataSetConfig, filterGuid);   
+            
+            displayPanel1.TestModeStartRequested += (sender, args) => { _testModeStart();  };
+            displayPanel1.TestModeStopRequested += (sender, args) => { _testModeStop(); };
+            displayPanel1.DisplayPanelValidatingError += (sender, args) =>
+            {
+                tabControlFsuipc.SelectedTab = displayTabPage;
+            };
         }
 
         private void initConfigRefDropDowns(DataSet dataSetConfig, string filterGuid)
@@ -71,7 +68,7 @@ namespace MobiFlight.UI.Dialogs
             dv.RowFilter = "guid <> '" + filterGuid + "'";
 
             configRefPanel.SetConfigRefsDataView(dv, filterGuid);
-            displayLedDisplayPanel.SetConfigRefsDataView(dv, filterGuid);
+            displayPanel1.SetConfigRefsDataView(dv, filterGuid);
             //displayShiftRegisterPanel.SetConfigRefsDataView(dv, filterGuid);
         }
 
@@ -104,7 +101,8 @@ namespace MobiFlight.UI.Dialogs
             }
             tabControlFsuipc.SelectedIndex = lastTabActive;
 
-            _initDisplayPanels();
+            // DISPLAY PANEL
+            displayPanel1.Init(_execManager);
 
             // PRECONDITION PANEL
             preconditionPanel.Init();
@@ -137,102 +135,7 @@ namespace MobiFlight.UI.Dialogs
             simConnectPanel1.LVars = (sender as List<String>);
         }
 
-        protected void _initDisplayPanels () {
-            // make all panels small and store the common height
-            groupBoxDisplaySettings.Controls.Add(displayPinPanel);
-            displayPinPanel.Dock = DockStyle.Top;
-            groupBoxDisplaySettings.Controls.Add(displayBcdPanel);
-            displayBcdPanel.Dock = DockStyle.Top;
-            groupBoxDisplaySettings.Controls.Add(displayLedDisplayPanel);
-            displayLedDisplayPanel.Dock = DockStyle.Top;
-            groupBoxDisplaySettings.Controls.Add(displayNothingSelectedPanel);
-            displayNothingSelectedPanel.Dock = DockStyle.Top;
-            groupBoxDisplaySettings.Controls.Add(servoPanel);
-            servoPanel.Dock = DockStyle.Top;
-            groupBoxDisplaySettings.Controls.Add(stepperPanel);
-            stepperPanel.Dock = DockStyle.Top;
-            groupBoxDisplaySettings.Controls.Add(displayShiftRegisterPanel);
-            displayShiftRegisterPanel.Dock = DockStyle.Top;
-            stepperPanel.OnManualCalibrationTriggered += new EventHandler<Panels.ManualCalibrationTriggeredEventArgs>(stepperPanel_OnManualCalibrationTriggered);
-            stepperPanel.OnSetZeroTriggered += new EventHandler(stepperPanel_OnSetZeroTriggered);
-            stepperPanel.OnStepperSelected +=  StepperPanel_OnStepperSelected;
-
-
-            groupBoxDisplaySettings.Controls.Add(displayLcdDisplayPanel);
-            displayLcdDisplayPanel.AutoSize = false;
-            displayLcdDisplayPanel.Height = 0;
-            displayLcdDisplayPanel.Dock = DockStyle.Top;
-
-            displayPanels.Clear();
-            displayPanelHeight = 0;
-            displayPanels.Add(displayPinPanel);
-            displayPanels.Add(displayBcdPanel);
-            displayPanels.Add(displayLedDisplayPanel);
-            displayPanels.Add(displayNothingSelectedPanel);
-            displayPanels.Add(servoPanel);
-            displayPanels.Add(stepperPanel);
-            displayPanels.Add(displayLcdDisplayPanel);
-            displayPanels.Add(displayShiftRegisterPanel);
-
-            foreach (UserControl p in displayPanels)
-            {
-                if (p.Height > 0 && (p.Height > displayPanelHeight)) displayPanelHeight = p.Height;                
-                p.Height = 0;
-            } //foreach
-
-            displayNothingSelectedPanel.Height = displayPanelHeight;
-            displayNothingSelectedPanel.Enabled = true;
-        }
-
-        private void StepperPanel_OnStepperSelected(object sender, EventArgs e)
-        {
-            // in case the module has changed
-            // we have to sync those changes to properly
-            // find the stepper and be able to
-            // show manual calibration group box or not.
-            display_syncToConfig();
-
-            String stepperAddress = (sender as ComboBox).SelectedValue.ToString();
-            String serial = SerialNumber.ExtractSerial(config.DisplaySerial);
-
-            MobiFlightStepper stepper = _execManager.getMobiFlightModuleCache()
-                                            .GetModuleBySerial(serial)
-                                            .GetStepper(stepperAddress);
-            stepperPanel.ShowManualCalibation(!stepper.HasAutoZero);
-            // sorry for this hack...
-        }
-
-        void stepperPanel_OnSetZeroTriggered(object sender, EventArgs e)
-        {
-            _syncFormToConfig();
-            String serial = SerialNumber.ExtractSerial(config.DisplaySerial);
-            _execManager.getMobiFlightModuleCache().resetStepper(serial, config.Stepper.Address);
-        }
-
-        void stepperPanel_OnManualCalibrationTriggered(object sender, Panels.ManualCalibrationTriggeredEventArgs e)
-        {
-            _syncFormToConfig();
-            int steps = e.Steps;
-            
-            String serial = SerialNumber.ExtractSerial(config.DisplaySerial);
-
-            MobiFlightStepper stepper = _execManager.getMobiFlightModuleCache()
-                                                    .GetModuleBySerial(serial)
-                                                    .GetStepper(config.Stepper.Address);
-
-            int CurrentValue = stepper.Position();
-            int NextValue = (CurrentValue + e.Steps);
-
-            _execManager.getMobiFlightModuleCache().setStepper(
-                serial,
-                config.Stepper.Address,
-                (NextValue).ToString(),
-                Int16.Parse(config.Stepper.OutputRev),
-                Int16.Parse(config.Stepper.OutputRev),
-                config.Stepper.CompassMode
-            );
-            
-        }        
+                
 
 #if ARCAZE
         /// <summary>
@@ -242,15 +145,13 @@ namespace MobiFlight.UI.Dialogs
         public void initWithArcazeCache (ArcazeCache arcazeCache)
         {
             List<ListItem> PreconditionModuleList = new List<ListItem>();
-            // update the display box with
-            // modules
-            displayModuleNameComboBox.Items.Clear();
-            displayModuleNameComboBox.Items.Add(new ListItem() { Value = "-", Label = "-" });
+            List<ListItem> DisplayModuleList = new List<ListItem>();
+
             
             foreach (IModuleInfo module in arcazeCache.getModuleInfo())
             {
                 arcazeFirmware[module.Serial] = module.Version;
-                displayModuleNameComboBox.Items.Add(new ListItem()
+                DisplayModuleList.Add(new ListItem()
                 {
                     Value = module.Name + "/ " + module.Serial,
                     Label = $"{module.Name} ({module.Serial})"
@@ -266,7 +167,7 @@ namespace MobiFlight.UI.Dialogs
 
             foreach (IModuleInfo module in _execManager.getMobiFlightModuleCache().getModuleInfo())
             {
-                displayModuleNameComboBox.Items.Add(new ListItem()
+                DisplayModuleList.Add(new ListItem()
                 {
                     Value = module.Name + "/ " + module.Serial,
                     Label = $"{module.Name} ({module.Port})"
@@ -276,10 +177,8 @@ namespace MobiFlight.UI.Dialogs
                 // preconditionPinSerialComboBox.Items.Add(module.Name + "/ " + module.Serial);
             }
 
-            displayModuleNameComboBox.ValueMember = "Value";
-            displayModuleNameComboBox.DisplayMember = "Label";
-            displayModuleNameComboBox.SelectedIndex = 0;
-
+            displayPanel1.SetArcazeSettings(arcazeFirmware, moduleSettings);
+            displayPanel1.SetModules(DisplayModuleList);
             preconditionPanel.SetModules(PreconditionModuleList);
         }
 #endif
@@ -289,15 +188,11 @@ namespace MobiFlight.UI.Dialogs
         /// </summary>
         public void initWithoutArcazeCache()
         {
-
-            // update the display box with
-            // modules
-            displayModuleNameComboBox.Items.Clear();
-            displayModuleNameComboBox.Items.Add(new ListItem() { Value = "-", Label = "-" });
-
+            List<ListItem> DisplayModuleList = new List<ListItem>();
+            
             foreach (IModuleInfo module in _execManager.getMobiFlightModuleCache().getModuleInfo())
             {
-                displayModuleNameComboBox.Items.Add(new ListItem()
+                DisplayModuleList.Add(new ListItem()
                 {
                     Value = module.Name + "/ " + module.Serial,
                     Label = $"{module.Name} ({module.Port})"
@@ -307,9 +202,7 @@ namespace MobiFlight.UI.Dialogs
                 // preconditionPinSerialComboBox.Items.Add(module.Name + "/ " + module.Serial);
             }
 
-            displayModuleNameComboBox.ValueMember = "Value";
-            displayModuleNameComboBox.DisplayMember = "Label";
-            displayModuleNameComboBox.SelectedIndex = 0;
+            displayPanel1.SetModules(DisplayModuleList);
         }
 #endif
         /// <summary>
@@ -319,45 +212,13 @@ namespace MobiFlight.UI.Dialogs
         /// <returns></returns>
         protected bool _syncConfigToForm(OutputConfigItem config)
         {
-            string serial = null;
             if (config == null) throw new Exception(i18n._tr("uiException_ConfigItemNotFound"));
 
             _syncFsuipcTabFromConfig(config);
 
             _syncComparisonTabFromConfig(config);
             
-            // third tab
-            if (!ComboBoxHelper.SetSelectedItem(displayTypeComboBox, config.DisplayType))
-            {
-                // TODO: provide error message
-                Log.Instance.log("_syncConfigToForm : Exception on selecting item in Display Type ComboBox", LogSeverity.Debug);
-            }
-
-            if (config.DisplaySerial != null && config.DisplaySerial != "")
-            {
-                serial = SerialNumber.ExtractSerial(config.DisplaySerial);
-                if (!ComboBoxHelper.SetSelectedItemByValue(displayModuleNameComboBox, config.DisplaySerial))
-                {
-                    // TODO: provide error message
-                }                
-            }
-
-
-            displayPinPanel.syncFromConfig(config);
-
-            displayBcdPanel.syncFromConfig(config);
-
-            displayLedDisplayPanel.syncFromConfig(config);
-
-            servoPanel.syncFromConfig(config);
-
-            // it is not nice but we haev to check what kind of stepper the stepper is
-            // to show or not show the manual calibration piece.
-            stepperPanel.syncFromConfig(config);
-
-            displayLcdDisplayPanel.syncFromConfig(config);
-
-            displayShiftRegisterPanel.SyncFromConfig(config);
+            displayPanel1.syncFromConfig(config);
 
             preconditionPanel.syncFromConfig(config);
             
@@ -419,25 +280,8 @@ namespace MobiFlight.UI.Dialogs
 
             config.Interpolation.Active = interpolationCheckBox.Checked;
             interpolationPanel1.syncToConfig(config.Interpolation);
-
-            display_syncToConfig();
-
+            displayPanel1.syncToConfig();
             preconditionPanel.syncToConfig(config);
-
-            // sync panels
-            displayPinPanel.syncToConfig(config);
-
-            displayLedDisplayPanel.syncToConfig(config);
-
-            displayBcdPanel.syncToConfig(config);
-
-            servoPanel.syncToConfig(config);
-
-            stepperPanel.syncToConfig(config);
-
-            displayLcdDisplayPanel.syncToConfig(config);
-
-            displayShiftRegisterPanel.SyncToConfig(config);
 
             return true;
         }
@@ -450,13 +294,6 @@ namespace MobiFlight.UI.Dialogs
             config.Comparison.Operand = comparisonOperandComboBox.Text;
             config.Comparison.IfValue = comparisonIfValueTextBox.Text;
             config.Comparison.ElseValue = comparisonElseValueTextBox.Text;
-        }
-
-        private void display_syncToConfig()
-        {
-            config.DisplayType = displayTypeComboBox.Text;
-            config.DisplayTrigger = "normal";
-            config.DisplaySerial = displayModuleNameComboBox.SelectedItem.ToString();
         }
 
         private void button1_Click(object sender, EventArgs e)
@@ -479,355 +316,6 @@ namespace MobiFlight.UI.Dialogs
         {
             _testModeStop();
             DialogResult = DialogResult.Cancel;
-        }
-
-        private void label1_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void displaySerialComboBox_SelectedIndexChanged(object sender, EventArgs e)
-        {
-#if DEBUG
-            Log.Instance.log("displaySerialComboBox_SelectedIndexChanged: called.", LogSeverity.Debug);
-#endif
-            // check which extension type is available to current serial
-            ComboBox cb = (sender as ComboBox);
-
-            try
-            {
-                // disable test button
-                // in case that no display is selected                
-                String serial = SerialNumber.ExtractSerial(cb.SelectedItem.ToString());
-
-                displayTypeComboBox.Enabled = groupBoxDisplaySettings.Enabled = testSettingsGroupBox.Enabled = (serial != "");
-                // serial is empty if no module is selected (on init of form)
-                //if (serial == "") return;                
-
-                // update the available types depending on the 
-                // type of module
-                if (serial.IndexOf("SN") != 0)
-                {
-                    displayTypeComboBox.Items.Clear();
-                    displayTypeComboBox.Items.Add("Pin");
-                    displayTypeComboBox.Items.Add(ArcazeLedDigit.TYPE);
-                    displayTypeComboBox.Items.Add(MobiFlightShiftRegister.TYPE);
-                    //displayTypeComboBox.Items.Add(ArcazeBcd4056.TYPE);
-                }
-                else
-                {
-                    displayTypeComboBox.Items.Clear();
-                    MobiFlightModule module = _execManager.getMobiFlightModuleCache().GetModuleBySerial(serial);
-                    foreach (DeviceType devType in module.GetConnectedOutputDeviceTypes())
-                    {
-#if DEBUG
-                        Log.Instance.log("displaySerialComboBox_SelectedIndexChanged: Adding Device Type: " + devType.ToString(), LogSeverity.Debug);
-#endif
-                        switch (devType)
-                        {
-                            case DeviceType.LedModule:
-                                displayTypeComboBox.Items.Add(ArcazeLedDigit.TYPE);
-                                break;
-
-                            case DeviceType.Output:
-                                displayTypeComboBox.Items.Add("Pin");
-                                //displayTypeComboBox.Items.Add(ArcazeBcd4056.TYPE);
-                                break;
-
-                            case DeviceType.Servo:
-                                displayTypeComboBox.Items.Add(DeviceType.Servo.ToString("F"));
-                                break;
-
-                            case DeviceType.Stepper:
-                                displayTypeComboBox.Items.Add(DeviceType.Stepper.ToString("F"));
-                                break;
-
-                            case DeviceType.LcdDisplay:
-                                displayTypeComboBox.Items.Add(DeviceType.LcdDisplay.ToString("F"));
-                                break;
-
-                            case DeviceType.ShiftRegister:
-                                displayTypeComboBox.Items.Add(MobiFlightShiftRegister.TYPE);
-                                break;
-                        }
-                    }
-
-                    if (displayTypeComboBox.Items.Count == 0)
-                    {
-                        if (MessageBox.Show(
-                                i18n._tr("uiMessageSelectedModuleDoesNotContainAnyOutputDevices"),
-                                i18n._tr("Hint"),
-                                MessageBoxButtons.YesNo) == System.Windows.Forms.DialogResult.Yes
-                            )
-                        {
-                            if (SettingsDialogRequested != null)
-                            {
-                                SettingsDialogRequested(module, new EventArgs());
-
-                                // trigger reload of Type ComboBox
-                                int CurrentIdx = displayModuleNameComboBox.SelectedIndex;
-                                displayModuleNameComboBox.SelectedIndex = 0;
-                                displayModuleNameComboBox.SelectedIndex = CurrentIdx;
-                            }
-                        }
-                    }
-                }
-
-                // third tab
-                if (!ComboBoxHelper.SetSelectedItem(displayTypeComboBox, config.DisplayType))
-                {
-                    // TODO: provide error message
-                    Log.Instance.log("displayArcazeSerialComboBox_SelectedIndexChanged : Problem setting Display Type ComboBox", LogSeverity.Debug);
-                }
-
-            }
-            catch (Exception ex)
-            {
-                displayPinPanel.displayPinBrightnessPanel.Visible = false;
-                displayPinPanel.displayPinBrightnessPanel.Enabled = false;
-                Log.Instance.log("displayArcazeSerialComboBox_SelectedIndexChanged : Some Exception occurred" + ex.Message, LogSeverity.Debug);
-            }
-        }
-
-        private void displayTypeComboBox_SelectedIndexChanged(object sender, EventArgs e)
-        {
-            Log.Instance.log("displayTypeComboBox_SelectedIndexChanged: called.", LogSeverity.Debug);
-
-            foreach (UserControl p in displayPanels)
-            {
-                p.Enabled = false;
-                p.AutoSize = false;
-                p.Height = 0;
-            } //foreach
-
-            try
-            {
-                bool panelEnabled = true;
-                // get the deviceinfo for the current arcaze
-                ComboBox cb = displayModuleNameComboBox;                
-                String serial = SerialNumber.ExtractSerial(cb.SelectedItem.ToString());
-
-                // we remove the callback method to ensure, that it is not added more than once
-                displayLedDisplayPanel.displayLedAddressComboBox.SelectedIndexChanged -= displayLedAddressComboBox_SelectedIndexChanged;
-
-#if ARCAZE
-                if (arcazeFirmware.ContainsKey(serial))
-                {
-
-                    switch ((sender as ComboBox).SelectedItem.ToString())
-                    {
-                        case "DisplayDriver":
-                            panelEnabled = ushort.Parse(arcazeFirmware[serial]) > 0x529;
-                            break;
-
-                        case "LedDriver2":
-                            panelEnabled = ushort.Parse(arcazeFirmware[serial]) > 0x554;
-                            break;
-
-                        case "LedDriver3":
-                            panelEnabled = ushort.Parse(arcazeFirmware[serial]) > 0x550;
-                            break;
-                    }
-
-                    displayPinPanel.displayPinBrightnessPanel.Visible = (moduleSettings[serial].type == SimpleSolutions.Usb.ArcazeCommand.ExtModuleType.LedDriver3);
-                    displayPinPanel.displayPinBrightnessPanel.Enabled = (displayPinPanel.displayPinBrightnessPanel.Visible && (cb.SelectedIndex > 1));
-                    
-                    //preconditionPortComboBox.Items.Clear();
-                    //preconditionPinComboBox.Items.Clear();
-
-                    List<ListItem> ports = new List<ListItem>();                    
-
-                    foreach (String v in ArcazeModule.getPorts())
-                    {
-                        ports.Add(new ListItem() { Label = v, Value = v });
-                        if (v == "B" || v == "E" || v == "H" || v == "K")
-                        {
-                            ports.Add(new ListItem() { Label = "-----", Value = "-----" });
-                        }
-
-                        if (v == "A" || v == "B")
-                        {
-                            //preconditionPortComboBox.Items.Add(v);
-                        }
-                    }
-
-                    displayPinPanel.SetPorts(ports);
-                    displayBcdPanel.SetPorts(ports);
-
-                    List<ListItem> pins = new List<ListItem>();
-                    foreach (String v in ArcazeModule.getPins())
-                    {
-                        pins.Add(new ListItem() { Label = v, Value = v });
-                        //preconditionPinComboBox.Items.Add(v);
-                    }
-
-                    displayPinPanel.SetPins(pins);
-                    displayBcdPanel.SetPins(pins);
-                    displayPinPanel.WideStyle = false;
-
-                    List<ListItem> addr = new List<ListItem>();
-                    List<ListItem> connectors = new List<ListItem>();
-                    foreach (string v in ArcazeModule.getDisplayAddresses()) addr.Add(new ListItem() { Label = v, Value = v });
-                    foreach (string v in ArcazeModule.getDisplayConnectors()) connectors.Add(new ListItem() { Label = v, Value = v });
-                    displayLedDisplayPanel.WideStyle = false;
-                    displayLedDisplayPanel.SetAddresses(addr);
-                    displayLedDisplayPanel.SetConnectors(connectors);                    
-                }
-                else 
-#endif
-                if (serial.IndexOf("SN") == 0)
-                {
-                    MobiFlightModule module = _execManager.getMobiFlightModuleCache().GetModuleBySerial(serial);
-
-                    displayPinPanel.SetModule(module);
-                    displayPinPanel.displayPinBrightnessPanel.Visible = true;
-                    displayPinPanel.displayPinBrightnessPanel.Enabled = (displayPinPanel.displayPinBrightnessPanel.Visible && (cb.SelectedIndex > 1));
-
-                    List<ListItem> outputs = new List<ListItem>();
-                    List<ListItem> ledSegments = new List<ListItem>();
-                    List<ListItem> servos = new List<ListItem>();
-                    List<ListItem> stepper = new List<ListItem>();
-                    List<ListItem> lcdDisplays = new List<ListItem>();
-                    List<ListItem> shiftRegisters = new List<ListItem>();
-
-
-                    foreach (IConnectedDevice device in module.GetConnectedDevices())
-                    {
-                        Log.Instance.log("displayTypeComboBox_SelectedIndexChanged: Adding connected device: " + device.Type.ToString() + ", " + device.Name, LogSeverity.Debug);
-                        switch (device.Type)
-                        {
-                            case DeviceType.LedModule:
-                                ledSegments.Add(new ListItem() { Value = device.Name, Label = device.Name });
-                                break;
-
-                            case DeviceType.Output:
-                                outputs.Add(new ListItem() { Value = device.Name, Label = device.Name });
-                                break;
-
-                            case DeviceType.Servo:
-                                servos.Add(new ListItem() { Value = device.Name, Label = device.Name });
-                                break;
-
-                            case DeviceType.Stepper:
-                                stepper.Add(new ListItem() { Value = device.Name, Label = device.Name });
-                                break;
-
-                            case DeviceType.LcdDisplay:
-                                int Cols = (device as MobiFlightLcdDisplay).Cols;
-                                int Lines= (device as MobiFlightLcdDisplay).Lines;
-                                lcdDisplays.Add(new ListItem() { Value = device.Name+","+ Cols+","+Lines, Label = device.Name });
-                                break;
-                            
-                            case DeviceType.ShiftRegister:
-                                shiftRegisters.Add(new ListItem() { Value = device.Name, Label = device.Name });                                
-                                
-                                break;
-                        }                        
-                    }
-                    displayPinPanel.WideStyle = true;
-                    displayPinPanel.SetPorts(new List<ListItem>());
-                    displayPinPanel.SetPins(outputs);
-
-                    displayLedDisplayPanel.WideStyle = true;
-                    displayLedDisplayPanel.displayLedAddressComboBox.SelectedIndexChanged += new EventHandler(displayLedAddressComboBox_SelectedIndexChanged);
-                    displayLedDisplayPanel.SetAddresses(ledSegments);
-
-                    servoPanel.SetAdresses(servos);
-
-                    stepperPanel.SetAdresses(stepper);
-
-                    displayShiftRegisterPanel.shiftRegistersComboBox.SelectedIndexChanged -= shiftRegistersComboBox_selectedIndexChanged;
-                    displayShiftRegisterPanel.shiftRegistersComboBox.SelectedIndexChanged += new EventHandler(shiftRegistersComboBox_selectedIndexChanged);
-                    displayShiftRegisterPanel.SetAddresses(shiftRegisters);
-
-                    displayLcdDisplayPanel.SetAddresses(lcdDisplays);
-                }
-                if ((sender as ComboBox).Text == "Pin")
-                {
-                    displayPinPanel.Enabled = panelEnabled;
-                    displayPinPanel.Height = displayPanelHeight;
-                }
-
-                if ((sender as ComboBox).Text == ArcazeBcd4056.TYPE)
-                {
-                    displayBcdPanel.Enabled = panelEnabled;
-                    displayBcdPanel.Height = displayPanelHeight;
-                }
-
-                if ((sender as ComboBox).Text == ArcazeLedDigit.TYPE)
-                {
-                    displayLedDisplayPanel.Enabled = panelEnabled;
-                    displayLedDisplayPanel.Height = displayPanelHeight;
-                }
-
-                if ((sender as ComboBox).Text == DeviceType.Servo.ToString("F"))
-                {
-                    servoPanel.Enabled = panelEnabled;
-                    servoPanel.Height = displayPanelHeight;
-                }
-
-                if ((sender as ComboBox).Text == DeviceType.Stepper.ToString("F"))
-                {
-                    stepperPanel.Enabled = panelEnabled;
-                    stepperPanel.Height = displayPanelHeight;
-                }
-                if ((sender as ComboBox).Text == DeviceType.LcdDisplay.ToString("F"))
-                {
-                    displayLcdDisplayPanel.Enabled = panelEnabled;
-                    displayLcdDisplayPanel.AutoSize = true;
-                    displayLcdDisplayPanel.Height = displayPanelHeight;
-                }
-
-                if ((sender as ComboBox).Text == DeviceType.ShiftRegister.ToString("F"))
-                {
-                    displayShiftRegisterPanel.Enabled = panelEnabled;
-                    displayShiftRegisterPanel.Height = displayPanelHeight;
-                }
-            }
-            catch (Exception exc)
-            {
-                Log.Instance.log("ConfigWizard.displayTypeComboBox_SelectedIndexChanged: EXC " + exc.Message, LogSeverity.Debug);
-                MessageBox.Show(i18n._tr("uiMessageNotImplementedYet"), 
-                                i18n._tr("Hint"), 
-                                MessageBoxButtons.OK, 
-                                MessageBoxIcon.Warning);
-            }
-        }
-
-        void displayLedAddressComboBox_SelectedIndexChanged(object sender, EventArgs e)
-        {
-            ComboBox cb = displayModuleNameComboBox;                
-            String serial = SerialNumber.ExtractSerial(cb.SelectedItem.ToString());
-            MobiFlightModule module = _execManager.getMobiFlightModuleCache().GetModuleBySerial(serial);
-
-            List<ListItem> connectors = new List<ListItem>();
-
-            foreach (IConnectedDevice device in module.GetConnectedDevices())
-            {
-                if (device.Type != DeviceType.LedModule) continue;
-                if (device.Name != ((sender as ComboBox).SelectedItem as ListItem).Value) continue;
-                for (int i = 0; i< (device as MobiFlightLedModule).SubModules; i++) {
-                    connectors.Add(new ListItem() { Label = (i + 1).ToString(), Value = (i + 1).ToString() });
-                }
-            }
-            displayLedDisplayPanel.SetConnectors(connectors);
-        }
-
-        private void shiftRegistersComboBox_selectedIndexChanged(object sender, EventArgs e)
-        {
-            ComboBox cb = displayModuleNameComboBox;
-            String serial = SerialNumber.ExtractSerial(cb.SelectedItem.ToString());
-            MobiFlightModule module = _execManager.getMobiFlightModuleCache().GetModuleBySerial(serial);
-            bool pwmSupport = false;
-
-            int numModules = 0;
-            foreach (IConnectedDevice device in module.GetConnectedDevices())
-            {
-                if (device.Type != DeviceType.ShiftRegister) continue;
-                if (device.Name != ((sender as ComboBox).SelectedItem as ListItem).Value) continue;
-                numModules = (device as MobiFlightShiftRegister).NumberOfShifters;
-            }
-            displayShiftRegisterPanel.SetNumModules(numModules);
         }
 
         private void fsuipcOffsetTextBox_Validating(object sender, CancelEventArgs e)
@@ -888,129 +376,7 @@ namespace MobiFlight.UI.Dialogs
             errorProvider.SetError(
                     control,
                     "");
-        }
-
-        private void displayArcazeSerialComboBox_Validating(object sender, CancelEventArgs e)
-        {
-            /* disabled this validation to permit configs even without module or
-             * as precondition only
-             
-            if (displayArcazeSerialComboBox.Text.Trim() == "-")
-            {
-                e.Cancel = true;
-                tabControlFsuipc.SelectedTab = displayTabPage;
-                displayArcazeSerialComboBox.Focus();
-                displayError(displayArcazeSerialComboBox, i18n._tr("uiMessageConfigWizard_SelectArcaze"));                
-            }
-            else
-            {
-               removeError(displayArcazeSerialComboBox);             
-            }
-             */
-        }
-
-        private void portComboBox_Validating(object sender, CancelEventArgs e)
-        {
-            ComboBox cb = (sender as ComboBox);
-            if (!cb.Parent.Visible) return;
-            if (null == cb.SelectedItem) return;
-            if (cb.SelectedItem.ToString() == "-----")
-            {
-                e.Cancel = true;
-                tabControlFsuipc.SelectedTab = displayTabPage;
-                cb.Focus();
-                displayError(cb, i18n._tr("Please_select_a_port"));
-            }
-            else
-            {
-                removeError(cb);
-            }
-        }
-
-        private void displayLedDisplayComboBox_Validating(object sender, CancelEventArgs e)
-        {
-            if (displayTypeComboBox.Text == ArcazeLedDigit.TYPE)                
-            {                
-                try
-                {
-                    int.Parse(displayLedDisplayPanel.displayLedAddressComboBox.Text);
-                    removeError(displayLedDisplayPanel.displayLedAddressComboBox);
-                }
-                catch (Exception exc)
-                {
-                    Log.Instance.log("displayLedDisplayComboBox_Validating : Parsing problem, " + exc.Message, LogSeverity.Debug);
-                    e.Cancel = true;
-                    tabControlFsuipc.SelectedTab = displayTabPage;
-                    displayLedDisplayPanel.displayLedAddressComboBox.Focus();
-                    displayError(displayLedDisplayPanel.displayLedAddressComboBox, i18n._tr("uiMessageConfigWizard_ProvideLedDisplayAddress"));
-                }                
-            }
-        }
-
-        private void comboBox1_SelectedIndexChanged(object sender, EventArgs e)
-        {
-            int value = Int16.Parse((sender as ComboBox).Text);
-            for (int i = 0; i < 8; i++)
-            {
-                displayLedDisplayPanel.displayLedDigitFlowLayoutPanel.Controls["displayLedDigit" + i + "CheckBox"].Visible = i < value;
-                displayLedDisplayPanel.displayLedDecimalPointFlowLayoutPanel.Controls["displayLedDecimalPoint" + i + "CheckBox"].Visible = i < value;
-                displayLedDisplayPanel.Controls["displayLedDisplayLabel" + i].Visible = i < value;
-
-                // uncheck all invisible checkboxes to ensure correct mask
-                if (i >= value)
-                {
-                    (displayLedDisplayPanel.displayLedDigitFlowLayoutPanel.Controls["displayLedDigit" + i + "CheckBox"] as CheckBox).Checked = false;
-                    (displayLedDisplayPanel.displayLedDecimalPointFlowLayoutPanel.Controls["displayLedDecimalPoint" + i + "CheckBox"] as CheckBox).Checked = false;
-                }
-            }
-        }
-
-        private void displayPinTestButton_Click(object sender, EventArgs e)
-        {
-            _testModeStart();
-        }
-
-        private void displayPinTestStopButton_Click(object sender, EventArgs e)
-        {
-            _testModeStop();
-        }
-
-        private void _testModeStart()
-        {
-            _syncFormToConfig();
-            displayPinTestStopButton.Enabled = true;
-            displayPinTestButton.Enabled = false;
-            displayTypeGroupBox.Enabled = false;
-            groupBoxDisplaySettings.Enabled = false;
-            try
-            {
-                _execManager.ExecuteTestOn(config);
-            }
-            catch (Exception e)
-            {
-                Log.Instance.log($"Error Test Mode execution. ExecuteTestOn > {e.Message}", LogSeverity.Error);
-            }
-        }
-
-        private void _testModeStop()
-        {
-            // check if running in test mode otherwise simply return
-            if (!displayPinTestStopButton.Enabled) return;
-
-            displayPinTestStopButton.Enabled = false;
-            displayPinTestButton.Enabled = true;
-            displayTypeGroupBox.Enabled = true;
-            groupBoxDisplaySettings.Enabled = true;
-            try
-            {
-                _execManager.ExecuteTestOff(config);
-            }
-            catch (Exception e)
-            {
-                Log.Instance.log($"Error Test Mode execution. ExecuteTestOff > {e.Message}", LogSeverity.Error);
-            }
-            
-        }
+        } 
 
         private void tabControlFsuipc_SelectedIndexChanged(object sender, EventArgs e)
         {
@@ -1019,40 +385,6 @@ namespace MobiFlight.UI.Dialogs
             _testModeStop();
         }
         
-        private void fsuipcOffsetTypeComboBox_SelectedIndexChanged(object sender, EventArgs e)
-        {
-            //updateByteSizeComboBox();
-        }
-
-        /*
-        private void updateByteSizeComboBox()
-        {
-            string selectedText = fsuipcSizeComboBox.Text;
-            fsuipcSizeComboBox.Items.Clear();
-            fsuipcSizeComboBox.Enabled = true;
-
-            if (fsuipcOffsetTypeComboBox.SelectedValue.ToString() == FSUIPCOffsetType.Integer.ToString())
-            {
-                fsuipcSizeComboBox.Items.Add("1");
-                fsuipcSizeComboBox.Items.Add("2");
-                fsuipcSizeComboBox.Items.Add("4");
-                fsuipcSizeComboBox.Items.Add("8");
-                ComboBoxHelper.SetSelectedItem(fsuipcSizeComboBox, selectedText);
-            }
-            else if (fsuipcOffsetTypeComboBox.SelectedValue.ToString() == FSUIPCOffsetType.Float.ToString())
-            {                
-                fsuipcSizeComboBox.Items.Add("4");
-                fsuipcSizeComboBox.Items.Add("8");
-                ComboBoxHelper.SetSelectedItem(fsuipcSizeComboBox, selectedText);
-            }
-            else if (fsuipcOffsetTypeComboBox.SelectedValue.ToString() == FSUIPCOffsetType.String.ToString())
-            {
-                fsuipcSizeComboBox.Enabled = false;
-            }
-        }
-        */
-
-
         private void interpolationCheckBox_CheckedChanged(object sender, EventArgs e)
         {
             interpolationPanel1.Enabled = (sender as CheckBox).Checked;
@@ -1070,6 +402,33 @@ namespace MobiFlight.UI.Dialogs
         private void ConfigWizard_FormClosing(object sender, FormClosingEventArgs e)
         {
             _execManager.GetSimConnectCache().LVarListUpdated -= ConfigWizard_LVarListUpdated;
+        }
+
+        private void _testModeStart()
+        {
+            _syncFormToConfig();
+
+            try
+            {
+                _execManager.ExecuteTestOn(config);
+            }
+            catch (Exception e)
+            {
+                Log.Instance.log($"Error Test Mode execution. ExecuteTestOn > {e.Message}", LogSeverity.Error);
+            }
+        }
+
+        private void _testModeStop()
+        {
+            try
+            {
+                _execManager.ExecuteTestOff(config);
+            }
+            catch (Exception e)
+            {
+                Log.Instance.log($"Error Test Mode execution. ExecuteTestOff > {e.Message}", LogSeverity.Error);
+            }
+
         }
     }
 }
